requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1
python-multipart>=0.0.6

# ML dependencies (required by moe_router)
torch>=2.0.0
transformers>=4.30.0
fasttext-wheel>=0.9.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# LoRA adapter support
peft>=0.7.0

# Quantization support (optional, for 4-bit models)
bitsandbytes>=0.41.0
accelerate>=0.24.0

# Utilities
requests>=2.31.0

# HTTP client — used by GatewayService for coordinator→worker dispatch
httpx>=0.27.0

# Testing
pytest>=7.0.0
